backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
numpy
scikit-learn
supabase
python-dotenv
httpx
